+++ Examples/SwiftUI/Superwall-SwiftUI/AppDelegate.swift
@@ -12,7 +12,7 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
   ) -> Bool {
-    SuperwallService.initSuperwall()
+    SuperwallService.initialize()
 
     Task {
       await StoreKitService.shared.loadSubscriptionState()