+++ Sources/SuperwallKit/Superwall/Analytics/Internal Tracking/Trackable Events/UserInitiatedEvents.swift
@@ -27,7 +27,7 @@ enum UserInitiatedEvent {
 
     func getSuperwallParameters() async -> [String: Any] { [:] }
   }
-
+  /*
   // TODO: Check over this
   // MARK: - To be deprecated/deleted
   struct PushNotification: TrackableUserInitiatedEvent {
@@ -53,5 +53,5 @@ enum UserInitiatedEvent {
     let pushNotificationId: String?
     let canImplicitlyTriggerPaywall = true
     var customParameters: [String: Any] = [:]
-  }
+  }*/
 }