+++ Sources/SuperwallKit/Superwall/Analytics/Superwall Event/SuperwallEventObjc.swift
@@ -9,7 +9,7 @@ import Foundation
 
 /// Objective-C-only analytical events that are automatically tracked by Superwall.
 ///
-/// These events are tracked internally by the SDK and sent to the delegate method ``SuperwallKit/SuperwallDelegateObjc/didTrackSuperwallEvent(_:params:)``.
+/// These events are tracked internally by the SDK and sent to the delegate method ``SuperwallKit/SuperwallDelegateObjc/didTrackSuperwallEventInfo(_:)``.
 @objc(SWKSuperwallEvent)
 public enum SuperwallEventObjc: Int {
   /// When the user is first seen in the app, regardless of whether the user is logged in or not.
@@ -41,7 +41,7 @@ public enum SuperwallEventObjc: Int {
   /// This event can be used to trigger a paywall. Just add the `deepLink_open` event to a campaign.
   case deepLink
 
-  /// When a trigger is fired.
+  /// When the tracked event matches an event added as a paywall trigger in a campaign.
   case triggerFire
 
   /// When a paywall is opened.
@@ -62,14 +62,17 @@ public enum SuperwallEventObjc: Int {
   /// When the user completes checkout in the payment sheet and any product was purchased.
   case transactionComplete
 
+  /// When the user successfully restores their purchases.
+  case transactionRestore
+
+  /// When a transaction takes > 5 seconds to show the payment sheet.
+  case transactionTimeout
+
   /// When the user successfully completes a transaction for a subscription product with no introductory offers.
   case subscriptionStart
 
   /// When the user successfully completes a transaction for a subscription product with an introductory offer.
   case freeTrialStart
-
-  /// When the user successfully restores their purchases.
-  case transactionRestore
 
   /// When the user attributes are set.
   case userAttributes