+++ Sources/SuperwallKit/Superwall/Config/Options/PaywallOptions.swift
@@ -30,7 +30,7 @@ public final class PaywallOptions: NSObject {
   /// Defines the messaging of the alert presented to the user when restoring a transaction fails.
   public var restoreFailed = RestoreFailed()
 
-  /// Pre-loads and caches trigger paywalls and products when you initialize the SDK via ``SuperwallKit/Superwall/configure(apiKey:delegate:options:)-7doe5``. Defaults to `true`.
+  /// Pre-loads and caches trigger paywalls and products when you initialize the SDK via ``SuperwallKit/Superwall/configure(apiKey:delegate:options:)-65jyx``. Defaults to `true`.
   ///
   /// Set this to `false` to load and cache paywalls and products in a just-in-time fashion.
   ///