+++ Sources/SuperwallKit/Superwall/Paywall Presentation/PublicPresentation.swift
@@ -58,7 +58,6 @@ public extension Superwall {
   /// - Parameters:
   ///   -  event: The name of the event you wish to track.
   ///   - params: Optional parameters you'd like to pass with your event. These can be referenced within the rules of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped.
-  ///   - on: The view controller to present the paywall on. Adds a new window to present on if `nil`. Defaults to `nil`.
   ///   - products: An optional ``PaywallProducts`` object whose products replace the remotely defined paywall products. Defauls to `nil`.
   ///   - ignoreSubscriptionStatus: Presents the paywall regardless of subscription status if `true`. Defaults to `false`.
   ///   - presentationStyleOverride: A `PaywallPresentationStyle` object that overrides the presentation style of the paywall set on the dashboard. Defaults to `.none`.
@@ -115,12 +114,12 @@ public extension Superwall {
   ///   -  event: The name of the event you wish to track.
   ///   - params: Optional parameters you'd like to pass with your event. These can be referenced within the rules of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped.
   ///   - paywallOverrides: An optional ``PaywallOverrides`` object whose parameters override the paywall defaults. Use this to override products, presentation style, and whether it ignores the subscription status. Defaults to `nil`.
-  ///   - paywallState: An optional callback that provides updates on the state of the paywall via a ``PaywallState`` object.
+  ///   - paywallHandler: An optional callback that provides updates on the state of the paywall via a ``PaywallState`` object.
   static func track(
     event: String,
     params: [String: Any]? = nil,
     paywallOverrides: PaywallOverrides? = nil,
-    paywallState: ((PaywallState) -> Void)? = nil
+    paywallHandler: ((PaywallState) -> Void)? = nil
   ) {
     trackCancellable = publisher(
       forEvent: event,
@@ -128,7 +127,7 @@ public extension Superwall {
       paywallOverrides: paywallOverrides
     )
     .sink { state in
-      paywallState?(state)
+      paywallHandler?(state)
     }
   }